Move the class inside.
