+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/LyricList/Rows/Edit/TimeTagBlueprintContainer.cs
@@ -53,7 +53,7 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.LyricList.Rows.Edit
         protected override SelectionBlueprint<TimeTag> CreateBlueprintFor(TimeTag item)
             => new TimeTagSelectionBlueprint(item);
 
-        protected class TimeTagSelectionHandler : BindableSelectionHandler<TimeTag>
+        protected class TimeTagSelectionHandler : BindableSelectionHandler
         {
             [Resolved]
             private ILyricTimeTagsChangeHandler lyricTimeTagsChangeHandler { get; set; }
